feat: Add support for additional sensor types
This commit adds support for the following sensor types:
- TVOC sensor
- NOX sensor
- CO2 sensor
- Humidity sensor
- PM25 sensor

For each sensor type, a new service is created with its associated characteristics and handlers. The handlers retrieve the sensor data through a query to Prometheus. The retrieved data is logged and returned as an integer value.

Note: Imported files and required modules are excluded from this commit message.